back-end/back-end-102/lesson-11/pyproject.toml
[project]
name = "lesson-11"
version = "0.1.0"
description = ""
authors = [
    {name = "Ansar",email = "issayevansar1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "multipart (>=1.2.1,<2.0.0)",
    "uuid (>=1.30,<2.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
